wallpapers: add all wallpapers options and add proper filling color
Each wallpaper must have a proper color the fill the rest of the interface, otherwise it's filled with blakc by default.
Wallpapers does not have the right size to fill entirely the interface, so an additional color must be used for that and which must be
closed to color in the border to be harmonized with it.